target-ppc: add CPU access_type into the migration stream
This is referenced in cpu_ppc_handle_mmu_fault() and so should be included
in the migration stream.

While we are here, change access_type from int to uint8_t as suggested by
Alexey.

+++ target-ppc/cpu.h
@@ -970,7 +970,7 @@ struct CPUPPCState {
     /* Next instruction pointer */
     target_ulong nip;
 
-    int access_type; /* when a memory exception occurs, the access
+    uint8_t access_type; /* when a memory exception occurs, the access
                         type is stored here */
 
     CPU_COMMON
